requirements.txt
boto==2.49.0
boto3==1.20.47
botocore==1.23.47
jmespath==0.10.0
mysql-connector-python==8.0.28
protobuf==3.19.4
python-dateutil==2.8.2
s3transfer==0.5.1
six==1.16.0
urllib3==1.26.8
